+++ core/core-common/src/core_clock.rs
@@ -1,4 +1,7 @@
-use std::time::{Duration, Instant};
+use std::{
+    sync::atomic::{AtomicBool, Ordering},
+    time::{Duration, Instant},
+};
 
 use parking_lot::Mutex;
 
@@ -30,7 +33,11 @@ impl CoreClock {
     }
 
     pub fn auto_lock_elapsed(&self) -> Duration {
-        self.auto_lock_accessed.elapsed()
+        if self.auto_lock_accessed.just_created.load(Ordering::Acquire) {
+            Duration::ZERO
+        } else {
+            self.auto_lock_accessed.elapsed()
+        }
     }
 
     pub fn pin_code_elapsed(&self) -> Duration {
@@ -43,6 +50,15 @@ impl CoreClock {
 
     pub fn pin_code_tick(&self) {
         self.pin_code_accessed.tick();
+        self.pin_code_accessed
+            .accessed
+            .store(true, Ordering::Release);
+    }
+
+    pub fn auto_lock_accessed(&self) {
+        self.auto_lock_accessed
+            .accessed
+            .store(true, Ordering::Release);
     }
 }
 
@@ -59,6 +75,8 @@ impl CoreClock {
 
 pub struct ActivityClock {
     last_activity: Mutex<Instant>,
+    accessed: AtomicBool,
+    just_created: AtomicBool,
 }
 
 impl ActivityClock {
@@ -66,11 +84,16 @@ impl ActivityClock {
     pub fn new(now: Instant) -> Self {
         Self {
             last_activity: Mutex::new(now),
+            accessed: AtomicBool::new(true),
+            just_created: AtomicBool::new(true),
         }
     }
 
     pub fn tick(&self) {
-        *self.last_activity.lock() = Instant::now();
+        if self.accessed.swap(false, Ordering::Acquire) {
+            *self.last_activity.lock() = Instant::now();
+            self.just_created.store(false, Ordering::Release);
+        }
     }
 
     pub fn elapsed(&self) -> Duration {