+++ mail/mail-common/src/models/conversation.rs
@@ -10,14 +10,15 @@ use crate::actions::{
     ConversationAction, ConversationAvailableActions, GeneralActions, LabelAsAction,
     MailActionError, MoveAction, MoveItemAction, filter_responses,
 };
+use crate::datatypes::dependencies::MessageOrConversationDependencyFetcher;
 use crate::datatypes::{
     AttachmentMetadata, ConversationLabelsCount, CustomLabel, Disposition, ExclusiveLocation,
     LocalMessageId, MessageAttachmentInfos, MessageLabelsCount, MessageRecipients, MessageSenders,
     ReadFilter, SystemLabelId,
 };
-use crate::find_in_query;
 use crate::models::*;
 use crate::{AppError, actions::conversations::Delete};
+use crate::{MailContextError, find_in_query};
 use anyhow::{Context, anyhow};
 use futures::future;
 use indoc::{formatdoc, indoc};
@@ -58,7 +59,7 @@ use std::collections::{BTreeSet, HashMap, HashSet};
 use std::future::Future;
 use std::ops::AddAssign;
 use std::sync::Arc;
-use tracing::{debug, error, info, trace, warn};
+use tracing::{debug, error, trace, warn};
 
 #[derive(Clone, Debug, Eq, Model, PartialEq, SmartDefault)]
 #[TableName("conversations")]
@@ -606,13 +607,13 @@ impl Conversation {
                 .query_values::<_, LocalMessageId>(
                     indoc::indoc! {"
                     WITH conv_msgs AS (
-                        SELECT local_id, ? AS label_id 
-                        FROM messages 
+                        SELECT local_id, ? AS label_id
+                        FROM messages
                         WHERE local_conversation_id=?
                     )
                     INSERT OR IGNORE INTO
                         message_labels (local_message_id, local_label_id)
-                    SELECT * FROM conv_msgs 
+                    SELECT * FROM conv_msgs
                     RETURNING local_message_id AS value
                     "},
                     params![label_id, id],
@@ -1305,7 +1306,7 @@ impl Conversation {
         }
 
         fn first_consecutive_unread_msg(
-            label_id: &LabelId,
+            label_id: Option<&LabelId>,
             messages: &[Message],
             filter: impl Fn(&Message) -> bool,
         ) -> Option<LocalMessageId> {
@@ -1323,7 +1324,9 @@ impl Conversation {
                 messages
                     .iter()
                     .rev()
-                    .find(|m| filter(m) && m.label_ids.contains(label_id))
+                    .find(|m| {
+                        filter(m) && label_id.is_none_or(|label_id| m.label_ids.contains(label_id))
+                    })
                     .and_then(|m| m.local_id)
             })
         }
@@ -1331,7 +1334,13 @@ impl Conversation {
         let view_is_starred_label_or_folder = label.label_type == LabelType::Label
             || label.label_type == LabelType::Folder
             || label.remote_id == Some(LabelId::starred());
-        let label_id = label.remote_id.as_ref()?;
+        // If this is not a custom label or a folder we don't want to match against
+        // label id.
+        let label_id = if label.label_type != LabelType::System {
+            Some(label.remote_id.as_ref()?)
+        } else {
+            None
+        };
 
         if view_is_starred_label_or_folder {
             first_consecutive_unread_msg(label_id, messages, |msg| !msg.flags.is_draft())
@@ -1660,10 +1669,12 @@ impl Conversation {
             };
             // Find all messages that need to be marked as read.
             let message = Message::find_first(
-                "WHERE local_conversation_id=?
+                "
+                JOIN message_labels AS ml ON messages.local_id = ml.local_message_id AND local_label_id=?
+                WHERE local_conversation_id=?
                 AND unread=0
                 ORDER BY time DESC",
-                params![conversation_id],
+                params![local_label_id, conversation_id],
                 bond,
             )
             .await?;
@@ -1799,7 +1810,7 @@ impl Conversation {
                     DELETE FROM message_labels
                     WHERE local_message_id IN (
                         SELECT local_id
-                        FROM messages 
+                        FROM messages
                         WHERE local_conversation_id=?1
                     ) AND message_labels.local_label_id=?2
                     RETURNING local_message_id AS value
@@ -1892,27 +1903,15 @@ impl Conversation {
         conversations: &[ApiConversation],
         api: &Proton,
         tether: &mut Tether,
-    ) -> Result<(), AppError> {
-        let mut missing_labels_ids = vec![];
-        for conv in conversations {
-            for label in &conv.labels {
-                let rid = label.id.clone();
-                if (Label::find_by_remote_id(rid, tether)).await?.is_none() {
-                    missing_labels_ids.push(label.id.clone());
-                }
-            }
+    ) -> Result<(), MailContextError> {
+        let mut fetcher = MessageOrConversationDependencyFetcher::new();
+
+        for conversation in conversations {
+            fetcher.check_api_conversation(conversation, tether).await?
         }
 
-        if !missing_labels_ids.is_empty() {
-            info!(
-                "{} label(s) were in a conversations but not locally, synchronizing...",
-                missing_labels_ids.len()
-            );
-            let missing_labels = Label::get_labels_by_ids(api, missing_labels_ids).await?;
-            tether
-                .tx(async |tx| Label::store_labels(tx, missing_labels).await)
-                .await?;
-        }
+        fetcher.fetch_and_store(api, tether).await?;
+
         Ok(())
     }
     /// Search for conversations.
@@ -1933,13 +1932,9 @@ impl Conversation {
         options: GetConversationsOptions,
         api: &Proton,
         tether: &mut Tether,
-    ) -> Result<Vec<Conversation>, AppError> {
+    ) -> Result<Vec<Conversation>, MailContextError> {
         // Fetch all the conversations from the API
-        let conversations = api
-            .get_conversations(options)
-            .await
-            .context("Error fetching the conversations from the API")?
-            .conversations;
+        let conversations = api.get_conversations(options).await?.conversations;
 
         Self::sync_dependencies(&conversations, api, tether).await?;
 
@@ -2062,8 +2057,8 @@ impl Conversation {
             .query_values(
                 formatdoc! {"
                 SELECT local_label_id AS value
-                FROM conversation_labels 
-                WHERE 
+                FROM conversation_labels
+                WHERE
                     local_conversation_id in ({})"
                     , placeholders(ids)
                 },
@@ -2561,12 +2556,13 @@ impl Conversation {
         tx: &mut impl RunTransaction,
         session: &Session,
     ) -> Result<(), AppError> {
-        let Some(conversation) = Self::find_by_id(local_conversation_id, tx.tether()).await? else {
+        let Some(mut conversation) = Self::find_by_id(local_conversation_id, tx.tether()).await?
+        else {
             return Err(AppError::ConversationNotFound(local_conversation_id));
         };
 
         if !conversation.has_messages {
-            let Some(rid) = conversation.remote_id else {
+            let Some(ref rid) = conversation.remote_id else {
                 return Err(AppError::ConversationHasNoRemoteId(local_conversation_id));
             };
             debug!("Syncing conversation messages");
@@ -2578,14 +2574,18 @@ impl Conversation {
                 )));
             }
 
-            let conversation_response = session.api().get_conversation(rid).await.map_err(|e| {
-                error!("failed to download conversation messages: {e:?}");
-                AppError::from(e)
-            })?;
+            let conversation_response =
+                session
+                    .api()
+                    .get_conversation(rid.clone())
+                    .await
+                    .map_err(|e| {
+                        error!("failed to download conversation messages: {e:?}");
+                        AppError::from(e)
+                    })?;
 
-            tx.run_tx::<_, _>(async |tx| {
+            tx.run_tx::<_, _>(async move |tx| {
                 let message_metadata: Vec<ApiMessageMetadata> = conversation_response.messages;
-                let mut new_conversation: Conversation = conversation_response.conversation.into();
 
                 Message::create_or_update_messages_from_metadata(message_metadata, tx)
                     .await
@@ -2594,11 +2594,8 @@ impl Conversation {
                         e
                     })?;
 
-                new_conversation.local_id = conversation.local_id;
-                new_conversation.row_id = conversation.row_id;
-                new_conversation.has_messages = true;
-
-                new_conversation.save(tx).await.map_err(|e| {
+                conversation.has_messages = true;
+                conversation.save(tx).await.map_err(|e| {
                     error!("Failed to write conversation: {e:?}");
                     e
                 })?;
@@ -3091,7 +3088,7 @@ impl ConversationMessageLabelStats {
                 JOIN message_labels AS ML ON
                     ML.local_message_id = messages.local_id AND
                     ML.local_label_id = ?
-                WHERE 
+                WHERE
                     messages.local_conversation_id = ? AND
                     messages.local_id IN ({})
             ",