+++ mail/mail-tui/src/app_model/mailbox/conversations.rs
@@ -208,6 +208,11 @@ impl ConversationsState {
         message: Message,
         mbox: &Mailbox,
     ) -> Command<Messages> {
+        if let Message::ConversationState(ConversationMessage::Refreshed(conversations)) = message {
+            self.conversations_refreshed(conversations);
+            return Command::None;
+        }
+
         match &mut self.messages {
             MessagesStatus::None => {
                 let Message::ConversationState(message) = message else {
@@ -252,7 +257,7 @@ impl ConversationsState {
                         Command::task(async move {
                             let paginator = paginator_clone.lock().await;
                             let has_more = paginator.has_more().await.unwrap();
-                            let total = paginator.total();
+                            let total = paginator.total().await.unwrap();
                             let seen = paginator.seen().await.unwrap();
                             Command::message(Messages::DisplayInfo(
                                 Some("Has more".to_owned()),